requirements.txt
croniter==6.0.0
flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
waitress==3.0.2
git+https://github.com/luketanti/nsz.git
PyYAML==6.0.3
requests==2.32.5
Pillow==12.1.0
unzip_http==0.6
watchdog==6.0.0
Werkzeug==3.1.5
zstandard==0.25.0
pycryptodome==3.23.0
geoip2==4.8.0
